+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Added `nstd_core_panic_with_msg`.
 ### `nstd.os`
 - Added `nstd_os_unix_shared_lib_handle`.
+### `nstd.string`
+- `nstd_string_pop` now returns `NSTDOptionalUnichar`.
 
 # 0.4.0
 ### `nstd`